% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RidgeMSE.R
\name{Find_ridge_best_split}
\alias{Find_ridge_best_split}
\title{Find_ridge_best_split}
\usage{
Find_ridge_best_split(feat, y, linear.idx, current.splitting.idx, lambda,
  update_A_inv = update_A_inv_R)
}
\arguments{
\item{feat}{all features}

\item{y}{outcome vector}

\item{linear.idx}{feature indexes where a Ridge Regression should be fitted}

\item{current.splitting.idx}{the index of the feature we are currently
testing for splits.}
}
\value{
It returns a vector of two doubles which determines what the best
splitting point is, and what it's MSE is.
  \item{\code{split_val_best}}{the best splitting value}
  \item{\code{MSE_best}}{the best MSE}
}
\description{
This is the fast impelementation of the best split function when
the leaf prediciton is coming from ridge regression.
}
\examples{
set.seed(309814)
n <- 1000
feat <- matrix(rnorm(4 * n), ncol = 4)
feat[2,] <- feat[which.min(feat[,2]),]
feat[3,] <- feat[which.min(feat[,2]),]
feat[4,] <- feat[which.min(feat[,2]),]
feat[5,] <- feat[which.min(feat[,2]),] - 1
y <- rnorm(n)
linear.idx <- 1:4
current.splitting.idx <- 2
lambda <- .2
Find_ridge_best_split_linear_MASS(feat = feat, y = y, linear.idx = linear.idx,
                                 current.splitting.idx = current.splitting.idx,
                                 lambda = lambda)
Find_ridge_best_split_slow(feat = feat, y = y, linear.idx = linear.idx,
                          current.splitting.idx = current.splitting.idx,
                          lambda = lambda)
Find_ridge_best_split(feat = feat, y = y, linear.idx = linear.idx,
                     current.splitting.idx = current.splitting.idx,
                     lambda = lambda)
# Note that fast split is not estimating the RSS or the MSE, but the RSS -
# sum(y^2).
}
