% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RidgeMSE.R
\name{Find_ridge_best_split_linear_MASS}
\alias{Find_ridge_best_split_linear_MASS}
\title{Find_ridge_best_split_linear_MASS}
\usage{
Find_ridge_best_split_linear_MASS(feat, y, linear.idx, current.splitting.idx,
  lambda)
}
\arguments{
\item{feat}{all features}

\item{y}{outcome vector}

\item{linear.idx}{feature indexes where a Ridge Regression should be fitted}

\item{current.splitting.idx}{the index of the feature we are currently
testing for splits.}
}
\value{
It returns a vector of two doubles which determines what the best
splitting point is, and what it's MSE is.
  \item{\code{split_val_best}}{the best splitting value}
  \item{\code{MSE_best}}{the best MSE}
}
\description{
Loops through all possible splits of feature
current.splitting.idx. This is just a test function which uses the ridge
regression canned ridge function of MASS
}
\examples{
set.seed(309814)
feat <- matrix(rnorm(80), ncol = 4)
y <- rnorm(20)
linear.idx <- 1:4
current.splitting.idx <- 2
lambda <- .2
Find_ridge_best_split_linear_MASS(
  feat = feat,
  y = y,
  current.splitting.idx = current.splitting.idx,
  linear.idx = linear.idx,
  lambda = lambda
)
}
