% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RidgeMSE.R
\name{Find_ridge_best_split_slow}
\alias{Find_ridge_best_split_slow}
\title{Find_ridge_best_split_slow}
\usage{
Find_ridge_best_split_slow(feat, y, linear.idx, current.splitting.idx, lambda)
}
\arguments{
\item{feat}{all features}

\item{y}{outcome vector}

\item{linear.idx}{feature indexes where a Ridge Regression should be fitted}

\item{current.splitting.idx}{the index of the feature we are currently
testing for splits.}
}
\value{
It returns a vector of two doubles which determines what the best
splitting point is, and what it's MSE is.
  \item{\code{split_val_best}}{the best splitting value}
  \item{\code{MSE_best}}{the best MSE}
}
\description{
This is the fast impelementation of the best split function when
the leaf prediciton is coming from ridge regression.
}
\examples{
set.seed(309814)
feat <- matrix(rnorm(80), ncol = 4)
y <- rnorm(20)
linear.idx <- 1:4
current.splitting.idx <- 2
lambda <- .2
Find_ridge_best_split_slow_linear_MASS(
  feat = feat,
  y = y,
  current.splitting.idx = current.splitting.idx,
  linear.idx = linear.idx,
  lambda = lambda
)
}
