% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_linearTree.R
\name{findSplittingPoint_linear}
\alias{findSplittingPoint_linear}
\alias{selectBestFeature,}
\alias{selectBestFeature-method}
\title{findSplittingPoint_linear}
\usage{
findSplittingPoint_linear(x, y, featureList,
  sampleIndex = list(averagingSampleIndex = 1:length(y), splittingSampleIndex
  = 1:length(y)), nodesize = list(splittingNodeSize = 5, averagingNodeSize =
  5), splitrule = "variance", categoricalFeatureCols = list())
}
\arguments{
\item{x}{feature data set}

\item{y}{dependent outcome}

\item{featureList}{}

\item{sampleIndex}{a list containing two integer vecotrs 
`averagingSampleIndex` and `splittingSampleIndex` which specify which of the 
features is used for splitting and which is used for estimating the beta in 
the leaves}

\item{nodesize}{a list containing `splittingNodeSize` and `averagingNodeSize`
which specify the nodesizes for the splitting and the aceraging sets.}

\item{splitrule}{= "variance",}

\item{categoricalFeatureCols}{= list()}
}
\value{
A list of two outputs: "splitFeature" is the best feature to split
in order to minimize the split loss, "splitValue" is its corresponding split
value.
}
\description{
Finds the best splitting value and feature to split on
}
