% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_honestRFTree.R
\name{honestRFTree_input_checker}
\alias{honestRFTree_input_checker}
\title{honstRFTree Input Checker}
\usage{
honestRFTree_input_checker(nodesize, sampleIndex)
}
\arguments{
\item{nodesize}{The minimum observations contained in terminal nodes. The
default value is 5.}

\item{sampleIndex}{A list of the index of observations that are used as
averaging dataset. The index are based on the original dataset `x` and `y`
from forest. Essentially, `x[sampleIndex]` generates the whole splitting
dataset.}
}
\description{
Checks that the input parameters for an honest RF make sense
}
