% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RidgeMSE.R
\name{update_A_inv_R}
\alias{update_A_inv_R}
\title{update_A_inv}
\usage{
update_A_inv_R(A_inv_current, new_x, leftnode)
}
\arguments{
\item{A_inv_current}{current A matrix}

\item{new_x}{new feature to be added}
}
\value{
the updated A matrix
}
\description{
This function is needed in the Ridge
}
\examples{
set.seed(309814)
feat <- matrix(rnorm(80), ncol = 4)
lambda <- .2
feat_current <- feat[-20, ]
feat_current_c <- cbind(feat_current, 1)
A_inv_current <- solve(t(feat_current_c) \%*\% feat_current_c +
                         lambda * diag(rep(1, ncol(feat_current_c))))
new_x <- feat[20, ]
update_A_inv(A_inv_current = A_inv_current, new_x = new_x)
feat_c <- cbind(feat, 1)
solve(t(feat_c) \%*\% feat_c + lambda * diag(rep(1, ncol(feat_c)))) # truth
}
